+++ tests/plugin/test_broken_plugin.py
@@ -1,25 +1,12 @@
-from textwrap import dedent
-
 import pytest
 
 
-def test_broken_command_path_plugin(runner, test_root_path, _install_plugin, caplog):
+def test_broken_command_path_plugin(runner, test_root_path, _install_plugin, snapshot):
     config_path = test_root_path / "test_data" / "configs" / "broken_plugin_config.toml"
 
     result = runner.invoke(["--config-file", config_path, "connection", "list"])
 
-    assert (
-        caplog.messages[0]
-        == "Cannot register plugin [broken_plugin]: Invalid command path [snow broken run]. Command group [broken] does not exist."
-    )
-    assert result.output == dedent(
-        """+---------------------------------------+
-| connection_name | parameters          |
-|-----------------+---------------------|
-| test            | {'account': 'test'} |
-+---------------------------------------+
-    """
-    )
+    assert result.output == snapshot
 
 
 @pytest.fixture(scope="module")