+++ tests/plugin/test_failing_plugin.py
@@ -1,27 +1,14 @@
-from textwrap import dedent
-
 import pytest
 
 
-def test_failing_plugin(runner, test_root_path, _install_plugin, caplog):
+def test_failing_plugin(runner, test_root_path, _install_plugin, snapshot):
     config_path = (
         test_root_path / "test_data" / "configs" / "failing_plugin_config.toml"
     )
 
     result = runner.invoke(["--config-file", config_path, "connection", "list"])
 
-    assert (
-        caplog.messages[0]
-        == "Cannot register plugin [failing_plugin]: Some error in plugin"
-    )
-    assert result.output == dedent(
-        """+---------------------------------------+
-| connection_name | parameters          |
-|-----------------+---------------------|
-| test            | {'account': 'test'} |
-+---------------------------------------+
-    """
-    )
+    assert result.output == snapshot
 
 
 @pytest.fixture(scope="module")